[PL] TM: Extract LocalAssemblerInterface.
This allows not to include the TMFEM.h in the process implementation.
